Move Spring support to the BotCommands-spring module
