+++ BotCommands-spring/src/main/kotlin/io/github/freya022/botcommands/autoconfigure/BotCommandsAutoConfiguration.kt
