+++ BotCommands-spring/src/main/kotlin/io/github/freya022/botcommands/internal/core/SpringBotCommandsInitializer.kt
