+++ BotCommands-spring/src/main/kotlin/io/github/freya022/botcommands/internal/core/annotations/InternalComponentScan.kt
