+++ BotCommands-spring/src/main/kotlin/io/github/freya022/botcommands/internal/core/service/SpringBotCommandsBootstrap.kt
