+++ BotCommands-spring/src/main/resources/META-INF/spring.factories
