+++ BotCommands-spring/src/test/kotlin/io/github/freya022/botcommands/spring/test/SpringMain.kt
