+++ BotCommands-spring/src/test/resources/application.properties
