+++ settings.gradle.kts
@@ -3,3 +3,4 @@ rootProject.name = "BotCommands"
 enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
 
 include(":spring-properties-processor")
+include(":BotCommands-spring")
