+++ src/main/kotlin/io/github/freya022/botcommands/internal/core/config/ConfigurationAnnotations.kt
@@ -5,7 +5,7 @@ import org.intellij.lang.annotations.Language
 @Target(AnnotationTarget.PROPERTY)
 @Retention(AnnotationRetention.SOURCE)
 @MustBeDocumented
-internal annotation class ConfigurationValue(
+annotation class ConfigurationValue(
     val path: String,
     val defaultValue: String = "",
     @Language("Java", prefix = "", suffix = " x = null;") val type: String = "java.lang.Byte",
@@ -14,7 +14,7 @@ internal annotation class ConfigurationValue(
 @Target(AnnotationTarget.PROPERTY)
 @Retention(AnnotationRetention.SOURCE)
 @MustBeDocumented
-internal annotation class DeprecatedValue(
+annotation class DeprecatedValue(
     val reason: String,
     val level: DeprecationLevel = DeprecationLevel.WARNING,
     val replacement: String = "",
@@ -27,4 +27,4 @@ internal annotation class DeprecatedValue(
 
 @Target(AnnotationTarget.PROPERTY)
 @Retention(AnnotationRetention.SOURCE)
-internal annotation class IgnoreDefaultValue
+annotation class IgnoreDefaultValue
